requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
django-autoslug==1.9.8
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.1
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
gunicorn==20.1.0
iniconfig==1.1.1
install==1.3.5
mccabe==0.7.0
packaging==21.3
phonenumberslite==8.13.1
pluggy==1.0.0
psycopg2==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
Unidecode==1.3.6
